Make rendered objects dynamic instead of creating dynamic copy.
Previously all objects available to the renderers created a dynamic copy
before rendering. This ensured any changes made by the renderers would
not be propagated back to the original object. That design decision is
no longer desirable so now the rendered objects are dynamics themselves
and passed directly to the renderers.