Teach tinySite to parse document dates from the path
Previously, tinySite would only find dates in the file name of a
document. This is fine but when you blog for a long time, you end up
with *a lot* of documents in a single folder. This change allows you to
now create folders for the year, month, day, hour, minute in the path
and/or filename. Much more flexible.